baselines/fogitj/Copolymer-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ipykernel==6.12.1
ipython==7.31.1
jupyter-client==6.1.12
jupyter_core==4.9.2
keras-tuner
keras==2.3.0
matplotlib==3.5.1
matplotlib-inline==0.1.2
numpy==1.18.5
pandas==1.2.4
protobuf==3.20.3
rdkit-pypi==2021.3.1.5
scipy==1.4.1
seaborn==0.11.2
scikit-learn
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
xlrd==1.2.0
